+++ Package.swift
@@ -42,12 +42,18 @@ let package = Package(
             targets: ["CDYelpFusionKit"])
     ],
     dependencies: [
-        .package(url: "https://github.com/Alamofire/Alamofire.git", "5.2.2")
-        .package(url: "https://github.com/tristanhimmelman/ObjectMapper", "4.2.0")
+        .package(
+            url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.4.3"))
     ],
     targets: [
         .target(
+            dependencies: [
+                .product(name: "Alamofire", package: "Alamofire")
+            ],
+            linkerSettings: [
+                .linkedFramework("UIKit", .when(platforms: [.iOS, .tvOS]))
+            ],
             name: "CDYelpFusionKit",
-            path: "CDYelpFusionKit")
+            path: "Source")
     ],
     swiftLanguageVersions: [.v5])