+++ Source/CDYelpAPIClient.swift
@@ -25,6 +25,12 @@
 //  THE SOFTWARE.
 //
 
+#if !os(OSX)
+    import UIKit
+#else
+    import Foundation
+#endif
+
 import Alamofire
 
 public class CDYelpAPIClient: NSObject {
@@ -32,7 +38,7 @@ public class CDYelpAPIClient: NSObject {
     private let apiKey: String!
     private lazy var manager: Alamofire.Session = {
         if let apiKey = self.apiKey,
-            apiKey.count > 0 {
+           apiKey.count > 0 {
             // Get the default headers
             var headers = HTTPHeaders.default
             // Add the Authorization header
@@ -73,7 +79,7 @@ public class CDYelpAPIClient: NSObject {
     ///
     public func isAuthenticated() -> Bool {
         if self.apiKey != nil,
-            self.apiKey.count > 0 {
+           self.apiKey.count > 0 {
             return true
         }
         return false
@@ -121,7 +127,7 @@ public class CDYelpAPIClient: NSObject {
                                  attributes: [CDYelpAttributeFilter]?,
                                  completion: @escaping (CDYelpSearchResponse?) -> Void) {
         assert((latitude != nil && longitude != nil) ||
-            (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API search endpoint.")
+                (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API search endpoint.")
         if let radius = radius {
             assert((radius > 0 && radius <= 40000), "The radius must be 40,000 meters or less to query the Yelp Fusion API search endpoint.")
         }
@@ -146,19 +152,22 @@ public class CDYelpAPIClient: NSObject {
                                                          openAt: openAt,
                                                          attributes: attributes)
 
-            self.manager.request(CDYelpRouter.search(parameters: parameters)).responseObject { (response: DataResponse<CDYelpSearchResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.search(parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpSearchResponse, AFError>) in
 
-                switch response.result {
-                case .success(let searchResponse):
-                    if let error = searchResponse.error {
-                        print("searchBusinesses(byTerm) error: ", error.description ?? "")
+                    switch response.result {
+                    case .success(let searchResponse):
+                        if let error = searchResponse.error {
+                            print("searchBusinesses(byTerm) error: ", error.description ?? "")
+                        }
+                        completion(searchResponse)
+                    case .failure(let error):
+                        print("searchBusinesses(byTerm) failure: ", error.localizedDescription)
+                        completion(nil)
                     }
-                    completion(searchResponse)
-                case .failure(let error):
-                    print("searchBusinesses(byTerm) failure: ", error.localizedDescription)
-                    completion(nil)
                 }
-            }
         }
     }
 
@@ -179,19 +188,22 @@ public class CDYelpAPIClient: NSObject {
 
             let parameters = Parameters.phoneParameters(withPhoneNumber: phoneNumber)
 
-            self.manager.request(CDYelpRouter.phone(parameters: parameters)).responseObject { (response: DataResponse<CDYelpSearchResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.phone(parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpSearchResponse, AFError>) in
 
-                switch response.result {
-                case .success(let searchResponse):
-                    if let error = searchResponse.error {
-                        print("searchBusinesses(byPhone) error: ", error.description ?? "")
+                    switch response.result {
+                    case .success(let searchResponse):
+                        if let error = searchResponse.error {
+                            print("searchBusinesses(byPhone) error: ", error.description ?? "")
+                        }
+                        completion(searchResponse)
+                    case .failure(let error):
+                        print("searchBusinesses(byPhone) failure: ", error.localizedDescription)
+                        completion(nil)
                     }
-                    completion(searchResponse)
-                case .failure(let error):
-                    print("searchBusinesses(byPhone) failure: ", error.localizedDescription)
-                    completion(nil)
                 }
-            }
         }
     }
 
@@ -214,7 +226,7 @@ public class CDYelpAPIClient: NSObject {
                                    completion: @escaping (CDYelpSearchResponse?) -> Void) {
         assert(type != nil, "A transaction type is required to query the Yelp Fusion API transactions endpoint.")
         assert((latitude != nil && longitude != nil) ||
-            (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API transactions endpoint.")
+                (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API transactions endpoint.")
 
         if self.isAuthenticated() == true {
 
@@ -222,20 +234,23 @@ public class CDYelpAPIClient: NSObject {
                                                                latitude: latitude,
                                                                longitude: longitude)
 
-            self.manager.request(CDYelpRouter.transactions(type: type.rawValue,
-                                                           parameters: parameters)).responseObject { (response: DataResponse<CDYelpSearchResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.transactions(type: type.rawValue,
+                                                   parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpSearchResponse, AFError>) in
 
-                switch response.result {
-                case .success(let searchResponse):
-                    if let error = searchResponse.error {
-                        print("searchTransactions(byType) error: ", error.description ?? "")
+                    switch response.result {
+                    case .success(let searchResponse):
+                        if let error = searchResponse.error {
+                            print("searchTransactions(byType) error: ", error.description ?? "")
+                        }
+                        completion(searchResponse)
+                    case .failure(let error):
+                        print("searchTransactions(byType) failure: ", error.localizedDescription)
+                        completion(nil)
                     }
-                    completion(searchResponse)
-                case .failure(let error):
-                    print("searchTransactions(byType) failure: ", error.localizedDescription)
-                    completion(nil)
                 }
-            }
         }
     }
 
@@ -257,17 +272,20 @@ public class CDYelpAPIClient: NSObject {
 
             let parameters = Parameters.businessParameters(withLocale: locale)
 
-            self.manager.request(CDYelpRouter.business(id: id,
-                                                       parameters: parameters)).responseObject { (response: DataResponse<CDYelpBusiness, AFError>) in
+            self.manager
+                .request(CDYelpRouter.business(id: id,
+                                               parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpBusiness, AFError>) in
 
-                switch response.result {
-                case .success(let business):
-                    completion(business)
-                case .failure(let error):
-                    print("fetchBusiness(byId) failure: ", error.localizedDescription)
-                    completion(nil)
+                    switch response.result {
+                    case .success(let business):
+                        completion(business)
+                    case .failure(let error):
+                        print("fetchBusiness(byId) failure: ", error.localizedDescription)
+                        completion(nil)
+                    }
                 }
-            }
         }
     }
 
@@ -350,19 +368,22 @@ public class CDYelpAPIClient: NSObject {
                                                           limit: limit,
                                                           matchThresholdType: matchThresholdType)
 
-            self.manager.request(CDYelpRouter.matches(parameters: parameters)).responseObject { (response: DataResponse<CDYelpSearchResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.matches(parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpSearchResponse, AFError>) in
 
-                switch response.result {
-                case .success(let searchResponse):
-                    if let error = searchResponse.error {
-                        print("searchBusinessMatches(byType) error: ", error.description ?? "")
+                    switch response.result {
+                    case .success(let searchResponse):
+                        if let error = searchResponse.error {
+                            print("searchBusinessMatches(byType) error: ", error.description ?? "")
+                        }
+                        completion(searchResponse)
+                    case .failure(let error):
+                        print("searchBusinessMatches(byType) failure: ", error.localizedDescription)
+                        completion(nil)
                     }
-                    completion(searchResponse)
-                case .failure(let error):
-                    print("searchBusinessMatches(byType) failure: ", error.localizedDescription)
-                    completion(nil)
                 }
-            }
         }
     }
 
@@ -384,21 +405,27 @@ public class CDYelpAPIClient: NSObject {
         if self.isAuthenticated() == true {
 
             let parameters = Parameters.reviewsParameters(withLocale: locale)
+            let decoder = JSONDecoder()
+            decoder.dateDecodingStrategy = .formatted(DateFormatter.reviews)
 
-            self.manager.request(CDYelpRouter.reviews(id: id,
-                                                      parameters: parameters)).responseObject { (response: DataResponse<CDYelpReviewsResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.reviews(id: id,
+                                              parameters: parameters))
+                .validate()
+                .responseDecodable(decoder: decoder,
+                                   completionHandler: { (response: DataResponse<CDYelpReviewsResponse, AFError>) in
+                                    switch response.result {
+                                    case .success(let reviewsResponse):
+                                        if let error = reviewsResponse.error {
+                                            print("fetchReviews(forBusinessId) error: ", error.description ?? "")
+                                        }
+                                        completion(reviewsResponse)
+                                    case .failure(let error):
+                                        print("fetchReviews(forBusinessId) failure: ", error.localizedDescription)
+                                        completion(nil)
+                                    }
+                                   })
 
-                switch response.result {
-                case .success(let reviewsResponse):
-                    if let error = reviewsResponse.error {
-                        print("fetchReviews(forBusinessId) error: ", error.description ?? "")
-                    }
-                    completion(reviewsResponse)
-                case .failure(let error):
-                    print("fetchReviews(forBusinessId) failure: ", error.localizedDescription)
-                    completion(nil)
-                }
-            }
         }
     }
 
@@ -420,8 +447,8 @@ public class CDYelpAPIClient: NSObject {
                                        locale: CDYelpLocale?,
                                        completion: @escaping (CDYelpAutoCompleteResponse?) -> Void) {
         assert((text != nil && text.count > 0) &&
-            latitude != nil &&
-            longitude != nil, "A search term, latitude, and longitude are required to query the Yelp Fusion API autocomplete endpoint.")
+                latitude != nil &&
+                longitude != nil, "A search term, latitude, and longitude are required to query the Yelp Fusion API autocomplete endpoint.")
 
         if self.isAuthenticated() == true {
 
@@ -430,19 +457,22 @@ public class CDYelpAPIClient: NSObject {
                                                                longitude: longitude,
                                                                locale: locale)
 
-            self.manager.request(CDYelpRouter.autocomplete(parameters: parameters)).responseObject { (response: DataResponse<CDYelpAutoCompleteResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.autocomplete(parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpAutoCompleteResponse, AFError>) in
 
-                switch response.result {
-                case .success(let autocompleteResponse):
-                    if let error = autocompleteResponse.error {
-                        print("autocompleteBusinesses(byText) error: ", error.description ?? "")
+                    switch response.result {
+                    case .success(let autocompleteResponse):
+                        if let error = autocompleteResponse.error {
+                            print("autocompleteBusinesses(byText) error: ", error.description ?? "")
+                        }
+                        completion(autocompleteResponse)
+                    case .failure(let error):
+                        print("autocompleteBusinesses(byText) failure: ", error.localizedDescription)
+                        completion(nil)
                     }
-                    completion(autocompleteResponse)
-                case .failure(let error):
-                    print("autocompleteBusinesses(byText) failure: ", error.localizedDescription)
-                    completion(nil)
                 }
-            }
         }
     }
 
@@ -466,18 +496,23 @@ public class CDYelpAPIClient: NSObject {
         if self.isAuthenticated() == true {
 
             let parameters = Parameters.eventParameters(withLocale: locale)
+            let decoder = JSONDecoder()
+            decoder.dateDecodingStrategy = .formatted(DateFormatter.events)
 
-            self.manager.request(CDYelpRouter.event(id: id,
-                                                    parameters: parameters)).responseObject { (response: DataResponse<CDYelpEvent, AFError>) in
-
-                switch response.result {
-                case .success(let event):
-                    completion(event)
-                case .failure(let error):
-                    print("fetchEvent(forId) failure: ", error.localizedDescription)
-                    completion(nil)
-                }
-            }
+            self.manager
+                .request(CDYelpRouter.event(id: id,
+                                            parameters: parameters))
+                .validate()
+                .responseDecodable(decoder: decoder,
+                                   completionHandler: { (response: DataResponse<CDYelpEvent, AFError>) in
+                                    switch response.result {
+                                    case .success(let event):
+                                        completion(event)
+                                    case .failure(let error):
+                                        print("fetchEvent(forId) failure: ", error.localizedDescription)
+                                        completion(nil)
+                                    }
+                                   })
         }
     }
 
@@ -540,20 +575,26 @@ public class CDYelpAPIClient: NSObject {
                                                          longitude: longitude,
                                                          radius: radius,
                                                          excludedEvents: excludedEvents)
+            let decoder = JSONDecoder()
+            decoder.dateDecodingStrategy = .formatted(DateFormatter.events)
 
-            self.manager.request(CDYelpRouter.events(parameters: parameters)).responseObject { (response: DataResponse<CDYelpEventsResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.events(parameters: parameters))
+                .validate()
+                .responseDecodable(decoder: decoder,
+                                   completionHandler: { (response: DataResponse<CDYelpEventsResponse, AFError>) in
+                                    switch response.result {
+                                    case .success(let eventsResponse):
+                                        if let error = eventsResponse.error {
+                                            print("searchEvents(byLocale) error: ", error.description ?? "")
+                                        }
+                                        completion(eventsResponse)
+                                    case .failure(let error):
+                                        print("searchEvents(byLocale) failure: ", error.localizedDescription)
+                                        completion(nil)
+                                    }
+                                   })
 
-                switch response.result {
-                case .success(let eventsResponse):
-                    if let error = eventsResponse.error {
-                        print("searchEvents(byLocale) error: ", error.description ?? "")
-                    }
-                    completion(eventsResponse)
-                case .failure(let error):
-                    print("searchEvents(byLocale) failure: ", error.localizedDescription)
-                    completion(nil)
-                }
-            }
         }
     }
 
@@ -575,7 +616,7 @@ public class CDYelpAPIClient: NSObject {
                                    longitude: Double?,
                                    completion: @escaping (CDYelpEvent?) -> Void) {
         assert((latitude != nil && longitude != nil) ||
-            (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API featured event endpoint.")
+                (location != nil), "Either a latitude and longitude or a location are required to query the Yelp Fusion API featured event endpoint.")
 
         if self.isAuthenticated() == true {
 
@@ -583,17 +624,22 @@ public class CDYelpAPIClient: NSObject {
                                                                 location: location,
                                                                 latitude: latitude,
                                                                 longitude: longitude)
+            let decoder = JSONDecoder()
+            decoder.dateDecodingStrategy = .formatted(DateFormatter.events)
 
-            self.manager.request(CDYelpRouter.featuredEvent(parameters: parameters)).responseObject { (response: DataResponse<CDYelpEvent, AFError>) in
-
-                switch response.result {
-                case .success(let event):
-                    completion(event)
-                case .failure(let error):
-                    print("fetchFeaturedEvent(forLocale) failure: ", error.localizedDescription)
-                    completion(nil)
-                }
-            }
+            self.manager
+                .request(CDYelpRouter.featuredEvent(parameters: parameters))
+                .validate()
+                .responseDecodable(decoder: decoder,
+                                   completionHandler: { (response: DataResponse<CDYelpEvent, AFError>) in
+                                    switch response.result {
+                                    case .success(let event):
+                                        completion(event)
+                                    case .failure(let error):
+                                        print("fetchFeaturedEvent(forLocale) failure: ", error.localizedDescription)
+                                        completion(nil)
+                                    }
+                                   })
         }
     }
 
@@ -614,16 +660,19 @@ public class CDYelpAPIClient: NSObject {
 
             let parameters = Parameters.categoriesParameters(withLocale: locale)
 
-            self.manager.request(CDYelpRouter.allCategories(parameters: parameters)).responseObject { (response: DataResponse<CDYelpCategoriesResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.allCategories(parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpCategoriesResponse, AFError>) in
 
-                switch response.result {
-                case .success(let event):
-                    completion(event)
-                case .failure(let error):
-                    print("fetchCategories(forLocale) alias: ", error.localizedDescription)
-                    completion(nil)
+                    switch response.result {
+                    case .success(let event):
+                        completion(event)
+                    case .failure(let error):
+                        print("fetchCategories(forLocale) alias: ", error.localizedDescription)
+                        completion(nil)
+                    }
                 }
-            }
         }
     }
 
@@ -645,17 +694,20 @@ public class CDYelpAPIClient: NSObject {
 
             let parameters = Parameters.categoriesParameters(withLocale: locale)
 
-            self.manager.request(CDYelpRouter.categoryDetails(alias: alias.rawValue,
-                                                              parameters: parameters)).responseObject { (response: DataResponse<CDYelpCategoryResponse, AFError>) in
+            self.manager
+                .request(CDYelpRouter.categoryDetails(alias: alias.rawValue,
+                                                      parameters: parameters))
+                .validate()
+                .responseDecodable { (response: DataResponse<CDYelpCategoryResponse, AFError>) in
 
-                switch response.result {
-                case .success(let event):
-                    completion(event)
-                case .failure(let error):
-                    print("fetchCategory(forAlias) alias: ", error.localizedDescription)
-                    completion(nil)
+                    switch response.result {
+                    case .success(let event):
+                        completion(event)
+                    case .failure(let error):
+                        print("fetchCategory(forAlias) alias: ", error.localizedDescription)
+                        completion(nil)
+                    }
                 }
-            }
         }
     }